reorder declarations within contracts
Following the Solidity style guide:

https://solidity.readthedocs.io/en/latest/style-guide.html#order-of-layout

This commit reorganizes declarations so that they are grouped together as
follows:

1. Types
2. State variables
3. Events
4. Functions

It also reorders a function in MintAndBurnAdmin.sol in accordance with

https://solidity.readthedocs.io/en/latest/style-guide.html#order-of-functions

so that the private function comes after the external functions.

This commit makes an exception in this reorganization for
ReserveDollarEternalStorage, which groups the accessor and mutator functions
together with their corresponding state variable, which feels like a more
natural way to think about that particular contract.