remove misleading output about coverage
We were previously getting a message about the amount of coverage of Go lines,
but that's not what we care about. We care about Solidity lines. The statement
about Go lines was entirely misleading.

Fix this by not invoking Go's notion of coverage at all, and just using our own.