fix `make cover`
A new solc version was released, which broke `make cover`.

The break happend because `make cover` depends on bindings generated by
`sol-compiler`, and `sol-compiler` auto-updates its version of solc online.

The new version of solc targets Constantinople, but the node we are running for
coverage is on Byzantium, and the version difference caused the failure.

This commit fixes the issue by targeting the Byzantium EVM.