cmd/poke: add `send-eth` command
We don't have any tool in our toolkit right now that is as convenient for
sending ETH as poke is for sending RSVD. This commit changes that, allowing us
to send ETH with `poke send-eth`.

`poke` may not be quite the right place for this. Sending ETH is sometimes
useful for interacting with the Reserve Dollar, but isn't strictly about
interacting with Reserve Dollar per se. But it seemed the best place out of the obvious
available options.